Avoid hashtable lookup for an item that has been fetched from disk.
When doWalkTapQueue processes items in backfilledItems (these come from
the disk walk of disk backfill or tap-bg-fetch of backfill visitor), it
obtains the latest copy of the item from hashtable. This is not
necessary and only creates overhead due to locks taken on the hashtable
which might contend with frontend load. In any case, disk backfill
stocks up flusher items and both methods set checkpoints, ensuring no
mutations are lost. This fix removes this check when unnecessary.

Signed-off-by: Abhishek Nanda <ananda@zynga.com>