SEG-11264: incorrect status for disk based backfill with multikv store
This commit sets up a way to be able to test the issue. Every kvstore
can be delayed by bf_disk_sleep_time seconds when enabled with
bf_disk_sleep_enable. With the dispatcher displaying the name of the
task under stats, it is possible to reproduce a scenario where a single
kvstore works slower than the rest. If backfill reports completion
before every kvstore is finished, it can be tracked.

Also added stats and documentation for all backfill related parameters.

Signed-off-by: Abhishek Nanda <ananda@zynga.com>