SEG-11002: backfill replication triggers huge ep_overhead in master resulting in oom kill
Added a configurable parameter bf_num_keys_threshold which triggers a
disk backfill whenever the number of items in a vbucket happen to be
above this threshold. This will enable forcing disk based backfill when
there is a large number of keys with very small blob sizes.

Performed a few cleanups and added stats for both thresholds. Got rid of
minimum limit for bf_resident_threshold so that memory backfill can be
forcefully triggered by setting the right values.

Signed-off-by: Abhishek Nanda <ananda@zynga.com>

Conflicts:

	backfill.cc